executed & trace moved
