+++ ethcore/src/executive.rs
@@ -24,6 +24,8 @@ use substate::*;
 use trace::{Trace, Tracer, NoopTracer, ExecutiveTracer};
 use crossbeam;
 
+pub use types::executed::{Executed, ExecutionResult};
+
 /// Max depth to avoid stack overflow (when it's reached we start a new thread with VM)
 /// TODO [todr] We probably need some more sophisticated calculations here (limit on my machine 132)
 /// Maybe something like here: `https://github.com/ethereum/libethereum/blob/4db169b8504f2b87f7d5a481819cfb959fc65f6c/libethereum/ExtVM.cpp`
@@ -44,45 +46,6 @@ pub struct TransactOptions {
 	/// Check transaction nonce before execution.
 	pub check_nonce: bool,
 }
-
-/// Transaction execution receipt.
-#[derive(Debug, PartialEq, Clone)]
-pub struct Executed {
-	/// Gas paid up front for execution of transaction.
-	pub gas: U256,
-
-	/// Gas used during execution of transaction.
-	pub gas_used: U256,
-
-	/// Gas refunded after the execution of transaction.
-	/// To get gas that was required up front, add `refunded` and `gas_used`.
-	pub refunded: U256,
-
-	/// Cumulative gas used in current block so far.
-	///
-	/// `cumulative_gas_used = gas_used(t0) + gas_used(t1) + ... gas_used(tn)`
-	///
-	/// where `tn` is current transaction.
-	pub cumulative_gas_used: U256,
-
-	/// Vector of logs generated by transaction.
-	pub logs: Vec<LogEntry>,
-
-	/// Addresses of contracts created during execution of transaction.
-	/// Ordered from earliest creation.
-	///
-	/// eg. sender creates contract A and A in constructor creates contract B
-	///
-	/// B creation ends first, and it will be the first element of the vector.
-	pub contracts_created: Vec<Address>,
-	/// Transaction output.
-	pub output: Bytes,
-	/// The trace of this transaction.
-	pub trace: Option<Trace>,
-}
-
-/// Transaction execution result.
-pub type ExecutionResult = Result<Executed, ExecutionError>;
 
 /// Transaction executor.
 pub struct Executive<'a> {
