+++ ethcore/src/trace/mod.rs
@@ -25,12 +25,12 @@ pub mod flat;
 mod import;
 mod localized;
 mod noop_tracer;
-pub mod trace;
 
+pub use types::trace_types::*;
 pub use self::block::BlockTraces;
 pub use self::config::{Config, Switch};
 pub use self::db::TraceDB;
-pub use self::trace::Trace;
+pub use types::trace_types::trace::Trace;
 pub use self::noop_tracer::NoopTracer;
 pub use self::executive_tracer::ExecutiveTracer;
 pub use types::trace_types::filter::{Filter, AddressesFilter};
