+++ ethcore/src/types/mod.rs
@@ -23,3 +23,4 @@ pub mod tree_route;
 pub mod blockchain_info;
 pub mod log_entry;
 pub mod trace_types;
+pub mod executed;
