+++ ethcore/src/types/trace_types/trace.rs
